requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Deprecated==1.2.14
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
executing==2.0.1
img2pdf==0.5.1
ipython==8.24.0
jedi==0.19.1
lxml==5.2.2
Markdown==3.6
matplotlib-inline==0.1.7
packaging==24.0
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pikepdf==8.15.1
pillow==10.3.0
prompt_toolkit==3.0.45
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
wrapt==1.16.0
